tests(slurm_ops): reorganize unit tests for slurm_ops into modules
Other changes:

* Add unit tests for new gres.conf editor on `SlurmctldManager`.
