+++ dev-requirements.txt
@@ -1,5 +1,5 @@
 # lib deps
-slurmutils ~= 0.9.0
+slurmutils ~= 0.10.0
 python-dotenv ~= 1.0.1
 pyyaml >= 6.0.2
 distro ~=1.9.0
